refactor(eve): fold endpoint status resolution into ModelEndpointFactory
resolveStatus() now lives on the endpoint objects next to createModel, so
an endpoint's credential story has one home: codexEndpoint owns the
~/.codex/auth.json read, aiGatewayEndpoint owns the gateway key/OIDC
probe, and resolve-model-endpoint-status shrinks to pure dispatch. The
standalone codex status module is gone; TStatusOptions on the factory
interface carries each endpoint's injectable test seams.
